philaudit/src/requirements.txt
camelot-py==0.9.0
charset-normalizer==3.0.1
cryptography==39.0.1
future==0.18.3
fuzzywuzzy==0.18.0
ghostscript==0.7
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
isodate==0.6.1
jedi==0.18.2
Levenshtein==0.20.9
looseversion==1.1.2
lxml==4.9.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.24.1
opencv-python==4.7.0.68
openpyxl==3.0.10
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathlib==1.0.1
pdf2image==1.16.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
prov==2.0.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.17
pydot==1.4.2
Pygments==2.15.1
PyPDF2==2.10.3
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
typing_extensions==4.5.0
aiofiles==23.1.0
anyio==3.6.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
ci-info==0.3.0
click==8.1.3
configobj==5.0.8
configparser==5.3.0
et-xmlfile==1.1.0
etelemetry==0.3.0
filelock==3.12.0
fitz==0.0.1.dev2
h11==0.14.0
itsdangerous==2.1.2
PyMuPDF==1.22.2
pyparsing==3.0.9
pyxnat==1.5
rapidfuzz==2.15.1
rdflib==6.3.2
requests==2.31.0
scipy==1.10.1
simplejson==3.19.1
sniffio==1.3.0
starlette==0.27.0
traitlets==5.9.0
traits==6.3.2
urllib3==2.0.2
uvicorn==0.22.0
wcwidth==0.2.6
zipp==3.15.0
